+++ packages/card/src/components/StatisticCard/demos/chart.tsx
@@ -4,14 +4,14 @@ import { StatisticCard } from '@ant-design/pro-components';
 export default () => {
   return (
     <StatisticCard
-      title="大盘趋势"
-      tooltip="大盘说明"
+      title="Market Trend"
+      tooltip="Market Description"
       style={{ maxWidth: 480 }}
       extra={<EllipsisOutlined />}
       chart={
         <img
           src="https://gw.alipayobjects.com/zos/alicdn/a-LN9RTYq/zhuzhuangtu.svg"
-          alt="柱状图"
+          alt="Bar Chart"
           width="100%"
         />
       }