+++ packages/table/src/components/EditableTable/index.tsx
@@ -525,8 +525,6 @@ function FieldEditableTable<
       style={{
         maxWidth: '100%',
       }}
-      {...props?.formItemProps}
-      name={props.name}
       shouldUpdate={(prev, next) => {
         const name = [props.name].flat(1) as string[];
         try {
@@ -537,6 +535,8 @@ function FieldEditableTable<
           return true;
         }
       }}
+      {...props?.formItemProps}
+      name={props.name}
     >
       <EditableTable<DataType, Params, ValueType>
         tableLayout="fixed"