+++ packages/components/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.8.6](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.8.5...@ant-design/pro-components@2.8.6) (2025-02-17)
+
+**Note:** Version bump only for package @ant-design/pro-components
+
 ## [2.8.5](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.8.4...@ant-design/pro-components@2.8.5) (2025-02-12)
 
 **Note:** Version bump only for package @ant-design/pro-components